Move Item definitions to a separate file
We probably set to add more item definitions, so, move these definitions to a separate file under project system folder.
Not that these are only needed for project system support but also for build and design time support even without
the aid of an external integration with IDEs.

We'll place it here for a while until we can come up with better location!