Add project and package reference item definitions
Use 'ProjectReference' and 'PackageReference' protocols to include projects and packages in the solution file.
We'll later introduce various references that'll convert to either a project or a package based on build configuration.

This was previously one of the best features in DNX build system that was scrapped. We are now in the process of bringing
the unified dependencies back to the MSBuild SDK projects along with more customizations with existing '~Reference' specifications.