Support Central Packages versioning
using the new 'Directory.Packages.props' file
that'll be imported after the 'Directory.Build.props'

This also means that, anything declared in build props can be used in packages props but not vice-versa.
Knowing MSBuild's evaluation, the properties in packages props can also be used in items declared in build props.
This is the only exception!