Support Packages props in other SDKs
Import the new 'Directory.Packages.props' file that'll be imported after the 'Directory.Build.props'.
This also means that, anything declared in build props can be used in packages props but not vice-versa.
Knowing MSBuild's evaluation, the properties in packages props can also be used in items declared in build props.
This is the only exception!

OTHER CHANGES:
- Use generic 'MSBuildToolsPath' instead of 'MSBuildToolsPath32'.
- Add 'NuGetToolsPath' to point to directory containing NuGet props/targets/tasks.
- Replace 'DevEnvDir' with 'MSBuild.GetVsInstallRoot()' as it is far more accurate.
- Remove 'MSBuildUseVisualStudioDirectoryLayout' property. This was used only when testing with VS IDE.