Exclude Binary Log files in default includes
Binary log (aka `.binlog`) files are structured binary-encoded log files produced with `msbuild -bl`
Since, they are log files and not part of the build system, we can safely exclude them in default includes.

Note that we're not excluding `.log` (_also produced by MSBuild loggers_) files which are text files.
Since, the `.log` extension can also be used for other files that could be resource or content in builds.