Correct the task name in Workflow targets
The task 'CreateWorkflowManifestResourceName' was meant to be a common task having no implementation.
Be ware that this target will fail since there is no task in that name. It was meant to be a placeholder
for other implementations to come through in custom project scenarios (Ex. IronPython, IronRuby project systems).

Note to Implementors:
 You'll have to implement the task similar in structure to the ones in the language specific targets
 and use the task in your own language targets file using `UsingTask` element.

You can also refer this target as a reference for your own implementations overriding this one.