Move source language detection to a new Target
If there's an unsupported project comes into play with Workflow,
the IDE features editing, validation and type inference would fail
since the source language detection fails or the user failed to use
the supported project types with supported source language.

Also, use 'SourceLanguageExtension' instead of '_LanguageExtension' and make it
overridable, so, users can support custom extensions or even hook custom languages.

Infer source language extension from the `DefaultLanguageSourceExtension` property
which most language targets support and provide a proper value for it. But the value
starts with a dot `.`, so, we have to trim and use it.