Make 'Target*Name' properties public
First, there's no concept of access level in MSBuild XML dialect.
Prefixing '_' with properties means it's just not recommended to users.

We now make 'Target*Name' properties public (removing '_' prefix)
based on internal feedback. It is shown that these properties are useful
in the context of heavily customized builds. They are not much different
from '*Identifier' properties but there are times when those values diverge.