Rename 'LangId' to 'SourceLanguageTag'
Since, the property 'LangID' is being used elsewhere in the older build targets as display language ID,
we cannot use it here as it overrides that property and all hell break loose.

Thus we rename the 'LangId' property to 'SourceLanguageTag' and the corresponding 'LangVersion' property to 'SourceLanguageVersion'.
The reason to introduce a new version property for languages is that the 'LangVersion' is directly connected to the Compiler
and is not flexible when parsing the version values as it only accepts a certain set of values embedded in the compiler.

To make the version property flexible and to allow resolving compiler packages through this, we introduce a new property
aptly named 'SourceLanguageVersion'. Currently it doesn't do anything except directly applies into 'LangVersion'.
Eventually we'll make it intelligent to resolve compiler packages and light-up language features.