Introduce Common user options
Solution/Project file User options was previously specified using Solution/Project file + '.user'. This is very useful
when IDEs have some customizations that they want to store but do not want to edit the Solution/Project file.

For the Core and Solution SDKs, we provided the same User options import. Since the previous method have already
been in use by the Legacy project system, we need a new User options import. Initially, the name for the file is
a combination of Project/Solution file name (without the file extension) along with the '.user' file extension.

Eventually, we'll provide ways to customize the name, location of the options file, if really needed!