Use Common SDK where applicable
We can use Common SDK with Legacy SDK instead of the Core SDK,
since, the purpose of the Common SDK to take all the common logic
and put it in a place where we can share them and version them.

Also, remove references to Core SDK where it's not needed!