Switch to file-scoped namespaces
C# 10 introduced file-scoped namespaces where you can define a namespace
for the whole file without having to surround them in brackets `{}`!

From the start of the declaration till the end of file, the namespace declared is used.
If there's a code before the namespace declaration, it'll be global or root namespace instead.